+++ service/src/main/java/org/whispersystems/textsecuregcm/entities/PhoneNumberIdentityKeyDistributionRequest.java
@@ -22,7 +22,7 @@ public record PhoneNumberIdentityKeyDistributionRequest(
     @JsonDeserialize(using = IdentityKeyAdapter.Deserializer.class)
     @Schema(description="the new identity key for this account's phone-number identity")
     IdentityKey pniIdentityKey,
-    
+
     @NotNull
     @Valid
     @ArraySchema(
@@ -32,26 +32,26 @@ public record PhoneNumberIdentityKeyDistributionRequest(
             Exactly one message must be supplied for each enabled device other than the sending (primary) device.
             """))
     List<@NotNull @Valid IncomingMessage> deviceMessages,
-    
+
     @NotNull
     @Valid
     @Schema(description="""
         A new signed elliptic-curve prekey for each enabled device on the account, including this one.
         Each must be accompanied by a valid signature from the new identity key in this request.""")
-    Map<Long, @NotNull @Valid ECSignedPreKey> devicePniSignedPrekeys,
-    
+    Map<Byte, @NotNull @Valid ECSignedPreKey> devicePniSignedPrekeys,
+
     @Schema(description="""
         A new signed post-quantum last-resort prekey for each enabled device on the account, including this one.
         May be absent, in which case the last resort PQ prekeys for each device will be deleted if any had been stored.
         If present, must contain one prekey per enabled device including this one.
         Prekeys for devices that did not previously have any post-quantum prekeys stored will be silently dropped.
         Each must be accompanied by a valid signature from the new identity key in this request.""")
-    @Valid Map<Long, @NotNull @Valid KEMSignedPreKey> devicePniPqLastResortPrekeys,
+    @Valid Map<Byte, @NotNull @Valid KEMSignedPreKey> devicePniPqLastResortPrekeys,
 
     @NotNull
     @Valid
     @Schema(description="The new registration ID to use for the phone-number identity of each device, including this one.")
-    Map<Long, Integer> pniRegistrationIds) {
+    Map<Byte, Integer> pniRegistrationIds) {
 
   @AssertTrue
   public boolean isSignatureValidOnEachSignedPreKey() {
